refactor(cost): own processing adapter
