+++ src/services/portfolio_transaction_processing_service/app/infrastructure/__init__.py
@@ -43,11 +43,6 @@ from .cost_calculation_workflow import (
     normalize_cost_event_code,
     normalize_cost_fee_amount,
 )
-from .cost_processing_adapter import (
-    CostEffectsStager,
-    CostProcessingCompatibilityAdapter,
-    PortfolioNotFoundError,
-)
 from .pipeline_stage_processing_adapter import PipelineStageProcessingAdapter
 from .position_processing_adapter import PositionHistoryProcessingAdapter
 from .prometheus_cost_basis_observability import (
@@ -79,8 +74,6 @@ __all__ = [
     "SqlAlchemyCashflowRulesRepository",
     "CashflowRuleSetVersion",
     "CashflowStageResult",
-    "CostProcessingCompatibilityAdapter",
-    "CostEffectsStager",
     "PROMETHEUS_CORPORATE_ACTION_RECONCILIATION_OBSERVER",
     "PrometheusCorporateActionReconciliationObserver",
     "CostCalculationWorkflow",
@@ -89,7 +82,6 @@ __all__ = [
     "OpenLotStateUpdateScope",
     "PositionHistoryProcessingAdapter",
     "PipelineStageProcessingAdapter",
-    "PortfolioNotFoundError",
     "LinkedCashLegError",
     "NoCashflowRuleError",
     "PROMETHEUS_COST_BASIS_CALCULATION_OBSERVER",
