+++ src/services/portfolio_transaction_processing_service/app/infrastructure/cost_calculation_workflow.py
@@ -73,7 +73,7 @@ from .booked_transaction_event_mapper import (
     to_transaction_event,
     with_booked_transaction_fields,
 )
-from .cost_basis import StagedCostEffects
+from .cost_basis.staged_effects import StagedCostEffects
 from .cost_metrics import COST_PROCESSING_EXECUTION_TOTAL, COST_PROCESSING_OPEN_LOTS_RESTORED
 from .fx_event_mapper import to_fx_contract_instrument_event
